Core: resolve background layers once for every backend
