+++ takumi-raster/src/background_drawing.rs
@@ -1,13 +1,14 @@
 use std::sync::Arc;
 
 use image::Rgba;
-use smallvec::{SmallVec, smallvec};
+use smallvec::SmallVec;
 use takumi_core::{
   geometry::{ComputedLayout as Layout, Point, Size},
-  paint::{
-    ConicGradientTile, GradientOverlayTile, LinearGradientTile, RadialGradientTile,
-    collect_repeat_tile_positions, collect_spaced_tile_positions, collect_stretched_tile_positions,
+  layout::background::{
+    LayerTileStyle, ResolveBackgroundLayerInput, ResolveBackgroundLayersInput,
+    background_origin_box, resolve_background_layer,
   },
+  paint::{ConicGradientTile, GradientOverlayTile, LinearGradientTile, RadialGradientTile},
 };
 use tiny_skia::{IntSize, Pixmap, PixmapMut, PremultipliedColorU8};
 
@@ -32,39 +33,6 @@ pub(crate) struct TileLayer {
 }
 
 pub(crate) type TileLayers = Vec<TileLayer>;
-
-#[derive(Clone, Copy)]
-pub(crate) struct LayerTileStyle {
-  pub pos: PositionValue,
-  pub size: BackgroundSize,
-  pub repeat: BackgroundRepeat,
-  pub blend_mode: BlendMode,
-}
-
-pub(crate) struct ResolveLayerTilesInput<'a> {
-  /// `background-origin` positioning area: position and `space`/`round` basis.
-  pub area: Size<u32>,
-  /// Painting area (border box) that `repeat` tiles across.
-  pub paint: Size<u32>,
-  /// Offset of the positioning area within the border box.
-  pub origin_offset: Point<i32>,
-  pub context: &'a RenderContext,
-}
-
-pub(crate) struct ResolveTileLayersInput<'a> {
-  pub images: &'a [BackgroundImage],
-  pub positions: &'a [PositionValue],
-  pub sizes: &'a [BackgroundSize],
-  pub repeats: &'a [BackgroundRepeat],
-  pub blend_modes: &'a [BlendMode],
-  pub context: &'a RenderContext,
-  /// `background-origin` positioning area.
-  pub area: Size<u32>,
-  /// Painting area (border box) that `repeat` tiles across.
-  pub paint: Size<u32>,
-  /// Offset of the positioning area within the border box.
-  pub origin_offset: Point<i32>,
-}
 
 fn should_rasterize_repeated_tile(
   tile: &BackgroundTile,
@@ -98,18 +66,6 @@ fn rasterize_tile(tile: BackgroundTile) -> Result<BackgroundTile> {
     return Ok(tile);
   };
   Ok(BackgroundTile::Pixmap(Arc::new(pixmap)))
-}
-
-fn resolve_intrinsic_size(image: &BackgroundImage, context: &RenderContext) -> IntrinsicSizing {
-  let BackgroundImage::Url(url) = image else {
-    return IntrinsicSizing::default();
-  };
-
-  let Ok(source) = resolve_image(url, context) else {
-    return IntrinsicSizing::default();
-  };
-
-  source.intrinsic_sizing().scale(&context.sizing)
 }
 
 pub(crate) fn rasterize_layers(
@@ -364,126 +320,6 @@ impl BackgroundTile {
 /// `position` and `space`/`round` distribution resolve against `area`; `repeat`
 /// tiles across `paint` (the painting/border box) so a repeating layer still
 /// covers the clip region when origin and clip differ.
-#[derive(Clone, Copy)]
-struct AxisArea {
-  area: u32,
-  paint: u32,
-  offset: i32,
-}
-
-/// Resolves tile origins on one axis, in border-box coordinates.
-fn resolve_axis_tiles(
-  repeat: BackgroundRepeatStyle,
-  pos: PositionValue,
-  tile_size: u32,
-  axis: AxisArea,
-  sizing: &SizingContext,
-  is_x: bool,
-) -> (SmallVec<[i32; 1]>, u32) {
-  let anchor = |tile: u32| {
-    axis.offset
-      + if is_x {
-        resolve_position_component_x(pos, tile, axis.area, sizing)
-      } else {
-        resolve_position_component_y(pos, tile, axis.area, sizing)
-      }
-  };
-  let shift = |mut positions: SmallVec<[i32; 1]>| {
-    if axis.offset != 0 {
-      positions.iter_mut().for_each(|x| *x += axis.offset);
-    }
-    positions
-  };
-
-  match repeat {
-    BackgroundRepeatStyle::Repeat => (
-      collect_repeat_tile_positions(axis.paint, tile_size, anchor(tile_size)),
-      tile_size,
-    ),
-    BackgroundRepeatStyle::NoRepeat => (smallvec![anchor(tile_size)], tile_size),
-    BackgroundRepeatStyle::Space => (
-      shift(collect_spaced_tile_positions(axis.area, tile_size)),
-      tile_size,
-    ),
-    BackgroundRepeatStyle::Round => {
-      let (positions, rounded) = collect_stretched_tile_positions(axis.area, tile_size);
-      (shift(positions), rounded)
-    }
-  }
-}
-
-fn resolve_auto_axis_from_intrinsic(
-  auto_axis: AutoBackgroundAxis,
-  intrinsic_ratio: Option<f32>,
-  fixed_size: u32,
-) -> Option<u32> {
-  let ratio = intrinsic_ratio?;
-  if ratio == 0.0 {
-    return Some(0);
-  }
-
-  let resolved = match auto_axis {
-    AutoBackgroundAxis::Width => fixed_size as f32 * ratio,
-    AutoBackgroundAxis::Height => fixed_size as f32 / ratio,
-  };
-
-  Some(resolved.round() as u32)
-}
-
-pub(crate) fn resolve_length_to_position_component(
-  length: Length,
-  available: i32,
-  sizing: &SizingContext,
-) -> i32 {
-  match length {
-    Length::Auto => available / 2,
-    _ => length.to_px(sizing, available as f32) as i32,
-  }
-}
-
-fn calculate_available_space(area_size: u32, tile_size: u32) -> i32 {
-  i32::try_from(area_size)
-    .unwrap_or(i32::MAX)
-    .saturating_sub_unsigned(tile_size)
-}
-
-pub(crate) fn resolve_position_component_x(
-  comp: PositionValue,
-  tile_w: u32,
-  area_w: u32,
-  sizing: &SizingContext,
-) -> i32 {
-  let available = calculate_available_space(area_w, tile_w);
-  match comp.0.x {
-    PositionComponent::KeywordX(PositionKeywordX::Left) => 0,
-    PositionComponent::KeywordX(PositionKeywordX::Center) => available / 2,
-    PositionComponent::KeywordX(PositionKeywordX::Right) => available,
-    PositionComponent::KeywordY(_) => available / 2,
-    PositionComponent::Length(length) => {
-      resolve_length_to_position_component(length, available, sizing)
-    }
-  }
-}
-
-pub(crate) fn resolve_position_component_y(
-  comp: PositionValue,
-  tile_h: u32,
-  area_h: u32,
-  sizing: &SizingContext,
-) -> i32 {
-  let available = calculate_available_space(area_h, tile_h);
-  match comp.0.y {
-    PositionComponent::KeywordY(PositionKeywordY::Top) => 0,
-    PositionComponent::KeywordY(PositionKeywordY::Center) => available / 2,
-    PositionComponent::KeywordY(PositionKeywordY::Bottom) => available,
-    PositionComponent::KeywordX(_) => available / 2,
-    PositionComponent::Length(length) => {
-      resolve_length_to_position_component(length, available, sizing)
-    }
-  }
-}
-
-/// Rasterize a single background image into a tile of the given size.
 pub(crate) fn render_tile(
   image: &BackgroundImage,
   tile_w: u32,
@@ -569,120 +405,19 @@ pub(crate) fn render_tile(
 }
 
 /// Resolve tile image, positions along X and Y for a background-like layer.
-pub(crate) fn resolve_layer_tiles(
+fn resolve_layer_tiles(
   image: &BackgroundImage,
   style: LayerTileStyle,
-  input: ResolveLayerTilesInput<'_>,
+  input: ResolveBackgroundLayerInput<'_>,
 ) -> Result<Option<TileLayer>> {
-  let resolved_size = style.size.resolve(
-    input.area,
-    &input.context.sizing,
-    resolve_intrinsic_size(image, input.context),
-  );
-
-  if resolved_size.width == 0 || resolved_size.height == 0 {
-    return Ok(None);
-  }
-
-  let axis_x = AxisArea {
-    area: input.area.width,
-    paint: input.paint.width,
-    offset: input.origin_offset.x,
-  };
-  let axis_y = AxisArea {
-    area: input.area.height,
-    paint: input.paint.height,
-    offset: input.origin_offset.y,
-  };
-
-  let (xs, ys, tile_w, tile_h) = match resolved_size.auto_axis {
-    Some(AutoBackgroundAxis::Width) => {
-      let (ys, tile_h) = resolve_axis_tiles(
-        style.repeat.1,
-        style.pos,
-        resolved_size.height,
-        axis_y,
-        &input.context.sizing,
-        false,
-      );
-      let tile_w = if style.repeat.1 == BackgroundRepeatStyle::Round {
-        resolve_auto_axis_from_intrinsic(
-          AutoBackgroundAxis::Width,
-          resolved_size.intrinsic_ratio,
-          tile_h,
-        )
-        .unwrap_or(resolved_size.width)
-      } else {
-        resolved_size.width
-      };
-      let (xs, tile_w) = resolve_axis_tiles(
-        style.repeat.0,
-        style.pos,
-        tile_w,
-        axis_x,
-        &input.context.sizing,
-        true,
-      );
-      (xs, ys, tile_w, tile_h)
-    }
-    Some(AutoBackgroundAxis::Height) => {
-      let (xs, tile_w) = resolve_axis_tiles(
-        style.repeat.0,
-        style.pos,
-        resolved_size.width,
-        axis_x,
-        &input.context.sizing,
-        true,
-      );
-      let tile_h = if style.repeat.0 == BackgroundRepeatStyle::Round {
-        resolve_auto_axis_from_intrinsic(
-          AutoBackgroundAxis::Height,
-          resolved_size.intrinsic_ratio,
-          tile_w,
-        )
-        .unwrap_or(resolved_size.height)
-      } else {
-        resolved_size.height
-      };
-      let (ys, tile_h) = resolve_axis_tiles(
-        style.repeat.1,
-        style.pos,
-        tile_h,
-        axis_y,
-        &input.context.sizing,
-        false,
-      );
-      (xs, ys, tile_w, tile_h)
-    }
-    None => {
-      let (xs, tile_w) = resolve_axis_tiles(
-        style.repeat.0,
-        style.pos,
-        resolved_size.width,
-        axis_x,
-        &input.context.sizing,
-        true,
-      );
-      let (ys, tile_h) = resolve_axis_tiles(
-        style.repeat.1,
-        style.pos,
-        resolved_size.height,
-        axis_y,
-        &input.context.sizing,
-        false,
-      );
-      (xs, ys, tile_w, tile_h)
-    }
-  };
-
-  if xs.is_empty() || ys.is_empty() {
-    return Ok(None);
-  }
-
-  let Some(tile) = render_tile(image, tile_w, tile_h, input.context)? else {
+  let context = input.context;
+  let Some(geometry) = resolve_background_layer(image, style, input) else {
     return Ok(None);
   };
-  let tile = if should_rasterize_repeated_tile(&tile, &xs, &ys) {
+  let Some(tile) = render_tile(image, geometry.tile_width, geometry.tile_height, context)? else {
+    return Ok(None);
+  };
+  let tile = if should_rasterize_repeated_tile(&tile, &geometry.xs, &geometry.ys) {
     rasterize_tile(tile)?
   } else {
     tile
@@ -690,13 +425,13 @@ pub(crate) fn resolve_layer_tiles(
 
   Ok(Some(TileLayer {
     tile,
-    xs,
-    ys,
-    blend_mode: style.blend_mode,
+    xs: geometry.xs,
+    ys: geometry.ys,
+    blend_mode: geometry.blend_mode,
   }))
 }
 
-pub(crate) fn resolve_tile_layers(input: ResolveTileLayersInput<'_>) -> Result<TileLayers> {
+pub(crate) fn resolve_tile_layers(input: ResolveBackgroundLayersInput<'_>) -> Result<TileLayers> {
   let last_position = input.positions.last().copied().unwrap_or_default();
   let last_size = input.sizes.last().copied().unwrap_or_default();
   let last_repeat = input.repeats.last().copied().unwrap_or_default();
@@ -714,7 +449,7 @@ pub(crate) fn resolve_tile_layers(input: ResolveTileLayersInput<'_>) -> Result<T
     results.push(resolve_layer_tiles(
       image,
       style,
-      ResolveLayerTilesInput {
+      ResolveBackgroundLayerInput {
         area: input.area,
         paint: input.paint,
         origin_offset: input.origin_offset,
@@ -735,7 +470,7 @@ pub(crate) fn create_mask(
   let mask_size = context.style.mask_size.as_ref();
   let mask_repeat = context.style.mask_repeat.as_ref();
 
-  let layers = resolve_tile_layers(ResolveTileLayersInput {
+  let layers = resolve_tile_layers(ResolveBackgroundLayersInput {
     images: mask_image,
     positions: mask_position,
     sizes: mask_size,
@@ -791,41 +526,6 @@ pub(crate) fn create_mask(
   )
 }
 
-struct OriginBox {
-  offset: Point<f32>,
-  size: Size<f32>,
-}
-
-fn background_origin_box(origin: BackgroundOrigin, layout: Layout) -> OriginBox {
-  let border = layout.border;
-  let padding = layout.padding;
-  let inset = |left: f32, right: f32, top: f32, bottom: f32| OriginBox {
-    offset: Point { x: left, y: top },
-    size: Size {
-      width: layout.size.width - left - right,
-      height: layout.size.height - top - bottom,
-    },
-  };
-
-  match origin {
-    BackgroundOrigin::BorderBox => OriginBox {
-      offset: Point { x: 0.0, y: 0.0 },
-      size: layout.size,
-    },
-    BackgroundOrigin::PaddingBox => inset(border.left, border.right, border.top, border.bottom),
-    BackgroundOrigin::ContentBox => inset(
-      border.left + padding.left,
-      border.right + padding.right,
-      border.top + padding.top,
-      border.bottom + padding.bottom,
-    ),
-    _ => OriginBox {
-      offset: Point { x: 0.0, y: 0.0 },
-      size: layout.size,
-    },
-  }
-}
-
 pub(crate) fn collect_background_layers(
   context: &RenderContext,
   layout: Layout,
@@ -836,7 +536,7 @@ pub(crate) fn collect_background_layers(
   // repeating layer covers the clip region when origin and clip differ.
   let origin = background_origin_box(context.style.background_origin, layout);
 
-  let mut layers = resolve_tile_layers(ResolveTileLayersInput {
+  let mut layers = resolve_tile_layers(ResolveBackgroundLayersInput {
     images: context.style.background_image.as_deref().unwrap_or(&[]),
     positions: &context.style.background_position,
     sizes: &context.style.background_size,
@@ -873,61 +573,4 @@ pub(crate) fn collect_background_layers(
   }
 
   Ok(layers)
-}
-
-#[cfg(test)]
-mod tests {
-  use super::{resolve_position_component_x, resolve_position_component_y};
-  use crate::{
-    style::{
-      Length, PositionComponent, PositionKeywordX, PositionKeywordY, PositionValue, SizingContext,
-      SpacePair,
-    },
-    viewport::Viewport,
-  };
-
-  fn test_sizing() -> SizingContext {
-    let viewport = Viewport::new((100, 100));
-    SizingContext::builder()
-      .viewport(viewport)
-      .font_size(viewport.font_size)
-      .line_height(0.0)
-      .build()
-  }
-
-  #[test]
-  fn oversized_background_keywords_resolve_to_negative_offsets() {
-    let sizing = test_sizing();
-    let position = PositionValue(SpacePair::from_pair(
-      PositionComponent::KeywordX(PositionKeywordX::Right),
-      PositionComponent::KeywordY(PositionKeywordY::Bottom),
-    ));
-
-    assert_eq!(
-      resolve_position_component_x(position, 150, 100, &sizing),
-      -50
-    );
-    assert_eq!(
-      resolve_position_component_y(position, 150, 100, &sizing),
-      -50
-    );
-  }
-
-  #[test]
-  fn oversized_background_percentages_use_signed_available_space() {
-    let sizing = test_sizing();
-    let position = PositionValue(SpacePair::from_pair(
-      PositionComponent::Length(Length::Percentage(25.0)),
-      PositionComponent::Length(Length::Percentage(75.0)),
-    ));
-
-    assert_eq!(
-      resolve_position_component_x(position, 140, 100, &sizing),
-      -10
-    );
-    assert_eq!(
-      resolve_position_component_y(position, 140, 100, &sizing),
-      -30
-    );
-  }
 }
